Fixed up tests for get.
